test(mix_tailwinds): rename shared test helpers
