+++ packages/mix_tailwinds/test/div_and_span_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mix/mix.dart';
 import 'package:mix_tailwinds/mix_tailwinds.dart';
 
-import 'div_and_span_test_helpers.dart';
+import 'tailwinds_test_helpers.dart';
 
 void main() {
   testWidgets('Div picks flex layout when flex token is present', (
